refactor(sql-gen): address Copilot review — consolidate dialect, seal trait, static emitter
Three fixes from PR #316 review:

1. Consolidate SqlDialect — moved the existing 5-variant enum (ClickHouse,
   PostgreSQL, DuckDB, MySQL, SQLite) from `server::models` to
   `sql_generator`, the natural home since both the SQL layer and
   `graph_catalog::schema_types` already depend on it independently of
   the HTTP API. `server::models` now re-exports it so the existing
   `crate::server::models::SqlDialect` import in `graph_catalog` keeps
   working. The duplicate `Dialect` enum introduced in 91ce1eb is gone.

2. Seal trait surface — `SqlEmitter` and `emitter_for` are now
   `pub(crate)`. Only `SqlDialect` and `generate_sql` are public. Trait
   methods can evolve in later phases without forming a semver
   commitment to external implementers.

3. No Box per call — `emitter_for` now returns `&'static dyn SqlEmitter`
   backed by a `static` singleton (emitters are zero-sized types).
   Removes the per-invocation heap allocation in `generate_sql`.

Also dropped the `dialect()` trait method that was unused at Phase 0.1
(would have triggered `dead_code` under CI's `-D warnings`); will be
added back in Phase 0.2 when a caller actually needs it.

Verified: `cargo build`, `cargo fmt --all`, `cargo clippy --all-targets
-- -D warnings` all clean; full test suite green (1344 lib + 279
integration + 36 misc, zero failures).
