+++ src/sql_generator/mod.rs
@@ -6,48 +6,82 @@
 //! `docs/design/DELTAGRAPH_PLAN.md`).
 //!
 //! ## Phase 0.1 status
-//! The trait is in place but every method currently delegates to the existing
-//! `clickhouse_query_generator` so behavior is unchanged. Subsequent phases
-//! will (a) move ClickHouse-specific logic behind the trait and (b) add a
-//! Databricks emitter alongside.
+//! The trait is in place but the only emitter (`ClickhouseEmitter`) currently
+//! delegates to the existing `clickhouse_query_generator` so behavior is
+//! unchanged. The trait surface is `pub(crate)` so it can evolve freely in
+//! subsequent phases without forming a semver commitment to external users —
+//! only `SqlDialect` and `generate_sql` are public.
 
 use crate::render_plan::RenderPlan;
+use serde::{Deserialize, Serialize};
 
-pub mod clickhouse;
+pub(crate) mod clickhouse;
 
-/// Target SQL dialect. New variants are added as emitters are implemented.
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, Default)]
-pub enum Dialect {
+/// SQL dialect for query generation.
+///
+/// Currently only `ClickHouse` is implemented; the other variants exist for
+/// API forward-compatibility (they will return `UnsupportedDialectError` at
+/// the executor / emitter boundary).
+///
+/// This type lives here — not in `server::models` — because the SQL layer and
+/// the schema layer (`graph_catalog::schema_types`) both need it independently
+/// of the HTTP API. `server::models` re-exports it for backward compatibility.
+#[derive(Debug, Default, Serialize, Deserialize, PartialEq, Eq, Hash, Clone, Copy)]
+#[serde(rename_all = "lowercase")]
+pub enum SqlDialect {
+    #[serde(rename = "clickhouse")]
     #[default]
     ClickHouse,
+
+    #[serde(rename = "postgresql")]
+    PostgreSQL,
+
+    #[serde(rename = "duckdb")]
+    DuckDB,
+
+    #[serde(rename = "mysql")]
+    MySQL,
+
+    #[serde(rename = "sqlite")]
+    SQLite,
     // Databricks,  // planned — see docs/design/DELTAGRAPH_PLAN.md
 }
 
-impl Dialect {
-    pub fn name(self) -> &'static str {
+impl SqlDialect {
+    /// String identifier for the dialect (lowercase).
+    pub fn as_str(&self) -> &'static str {
         match self {
-            Dialect::ClickHouse => "clickhouse",
+            SqlDialect::ClickHouse => "clickhouse",
+            SqlDialect::PostgreSQL => "postgresql",
+            SqlDialect::DuckDB => "duckdb",
+            SqlDialect::MySQL => "mysql",
+            SqlDialect::SQLite => "sqlite",
         }
+    }
+
+    /// Whether an emitter is implemented for this dialect today.
+    pub fn is_supported(&self) -> bool {
+        matches!(self, SqlDialect::ClickHouse)
     }
 }
 
 /// Renders a `RenderPlan` into SQL text for a target dialect.
 ///
-/// During Phase 0.1 there is exactly one implementor (`ClickhouseEmitter`)
-/// that forwards to `clickhouse_query_generator::generate_sql`. Future phases
-/// will widen this surface area (function mapping, recursive CTE shape,
-/// quote rules, NULL semantics) so the trait can host real per-dialect logic.
-pub trait SqlEmitter: Send + Sync {
-    fn dialect(&self) -> Dialect;
-
-    /// Render a full `RenderPlan` into a SQL string.
+/// `pub(crate)` because the method surface will widen in later phases
+/// (function mapping, recursive CTE shape, quote rules, NULL semantics).
+/// External consumers should call the free function [`generate_sql`] instead.
+pub(crate) trait SqlEmitter: Send + Sync {
     fn emit(&self, plan: RenderPlan, max_cte_depth: u32) -> String;
 }
 
-/// Returns the emitter for a given dialect.
-pub fn emitter_for(dialect: Dialect) -> Box<dyn SqlEmitter> {
+/// Returns a borrowed static emitter for the given dialect.
+///
+/// No heap allocation: each emitter is a zero-sized type held in a `static`.
+pub(crate) fn emitter_for(dialect: SqlDialect) -> &'static dyn SqlEmitter {
+    static CLICKHOUSE: clickhouse::ClickhouseEmitter = clickhouse::ClickhouseEmitter;
     match dialect {
-        Dialect::ClickHouse => Box::new(clickhouse::ClickhouseEmitter),
+        SqlDialect::ClickHouse => &CLICKHOUSE,
+        d => unimplemented!("SQL emitter for dialect {:?} is not yet implemented", d),
     }
 }
 
@@ -55,7 +89,7 @@ pub fn emitter_for(dialect: Dialect) -> Box<dyn SqlEmitter> {
 ///
 /// Existing call sites in `clickhouse_query_generator::generate_sql` continue
 /// to work unchanged; new call sites should prefer this entry point so the
-/// dialect can be chosen at runtime.
+/// dialect can be chosen at runtime in a future phase.
 pub fn generate_sql(plan: RenderPlan, max_cte_depth: u32) -> String {
-    emitter_for(Dialect::default()).emit(plan, max_cte_depth)
+    emitter_for(SqlDialect::default()).emit(plan, max_cte_depth)
 }
